requirements.txt
boto3==1.23.10
botocore==1.26.10
click==8.1.3
colorama==0.4.5
Flask==2.2.2
Flask-Cors==3.0.10
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
numpy==1.23.3
opencv-python-headless==4.5.5.64
pandas==1.4.4
python-dateutil==2.8.2
pytz==2022.2.1
s3transfer==0.5.2
six==1.16.0
urllib3==1.26.12
Werkzeug==2.2.2
wincertstore==0.2
azure-storage-blob
